move tests to a dedicated file; pin function-body-mention behaviour
The '// @bun' tests in runtime-transpiler.test.ts fail on a second
release-build run when the entry module is served from the runtime
transpiler cache (the has_loaded flag is not set on that path; #33371
tracks it). Keep that file untouched and put the class-hoist TDZ tests
in a sibling file so they can be verified independently.

Also document that the use_count_estimate guard counts textual mentions
in already-visited statements (including function bodies), and extend
the 'still hoisted' test to pin that a class mentioned by a preceding
function body is left in source order while an unmentioned class is
still moved ahead.
